% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{attribute_pattern_table_header}
\alias{attribute_pattern_table_header}
\title{Generate attribute pattern table header}
\usage{
attribute_pattern_table_header(k, m = 2, order = k)
}
\arguments{
\item{k}{Number of Attributes.}

\item{m}{Number of Categories. Default 2 or dichotomous response.}

\item{order}{Order of the table. Default \code{k} or the full order.}
}
\value{
Return a matrix containing the class table
}
\description{
Generate attribute pattern table header
}
\examples{
# K = 3
attribute_pattern_table_header(3)

# K = 4
attribute_pattern_table_header(4)
}
