% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap.slcm}
\alias{bootstrap.slcm}
\title{Goodness of Fit Test with Bootstrap Resampling}
\usage{
bootstrap.slcm(object, method, B = 100, max.iter = 500, verbose = FALSE)
}
\arguments{
\item{object}{a slcm object}

\item{method}{estimation method for estimating parameters}

\item{B}{number of bootstrap resamples}

\item{max.iter}{maximum of iteration}

\item{verbose}{print}
}
\description{
Goodness of Fit Test with Bootstrap Resampling
}
