% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{compare}
\alias{compare}
\title{Comparing Two Estimated \code{slcm} Models}
\usage{
compare(
  model1,
  model2,
  test = c("chisq", "boot"),
  nboot = 100,
  maxiter = 500,
  tol = 1e-05,
  verbose = FALSE
)
}
\arguments{
\item{model1}{an object of class \code{slcm} and \code{estimated}.}

\item{model2}{another object of class \code{slcm} and \code{estimated}.}

\item{test}{a character string specifying the type of test to be conducted. If "chisq", a chi-squared test is conducted. If "boot", a bootstrap test is conducted.}

\item{nboot}{an integer specifying the number of bootstrap rounds to be performed.}

\item{maxiter}{an integer specifying maximum number of iterations allowed for the estimation process of each bootstrapping round.}

\item{tol}{a numeric value setting tolerance for the convergence of each bootstrapping round.}

\item{verbose}{a logical value indicating whether to print progress updates on the number of bootstrapping rounds completed.}
}
\value{
A \code{data.frame} containing the number of parameters (Df), loglikelihood, AIC, BIC, G-squared statistics, and the residual degree of freedom for each object.
Depending on the \code{test} argument, the p-value for the corresponding statistical test may also be included.
}
\description{
Provides relative model fit test for two fitted SLCM models with deviance statistic.
}
\seealso{
\link[slcm]{compare}
}
