% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{confint.slcm}
\alias{confint.slcm}
\title{Confidence Intervals for Model Parameters}
\usage{
\method{confint}{slcm}(
  object,
  parm,
  level = 0.95,
  method = c("asymp", "logit"),
  out = c("param", "logit")
)
}
\arguments{
\item{object}{a fitted slcm object.}

\item{parm}{a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{the confidence level required.}

\item{method}{calculating method for standard error}

\item{out}{the type of output}
}
\description{
Computes confidence intervals for one or more parameters of fitted model. Package \pkg{slcm} adds methods for \code{slcm} fits.
}
