% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage}
\alias{coverage}
\title{Coverage Probabilities for Parameter Estimates for \code{slcm} model}
\usage{
coverage(object, nsim = 100, parm, level = 0.95, verbose = TRUE, control, ...)
}
\arguments{
\item{object}{a \code{slcm} object}

\item{nsim}{number of simulation sample}

\item{parm}{parameter to be returned}

\item{level}{significnace level}

\item{verbose}{print}

\item{control}{slcm control}

\item{nobs}{number of observations for}
}
\description{
Coverage Probabilities for Parameter Estimates for \code{slcm} model
}
