% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\alias{estimate.slcm}
\title{Estimation for Parameters of \code{slcm} Object}
\usage{
estimate(object, ...)

\method{estimate}{slcm}(object, data,
    method = c("em", "hybrid", "nlm"),
    fix2zero = NULL,
    control = slcmControl(), ...)
}
\arguments{
\item{object}{a \code{slcm} object which defines latent structure to be estimated.}

\item{data}{a \code{data.frame} object which contains observed categorical variables beloning to the latent structure.}

\item{method}{estimating method for slcm parameters.}

\item{fix2zero}{a \code{vector} of parameters to be restricted to zero. The details of restriction is given under 'Details'}

\item{control}{a \code{list} of control for the estimation procedure to change default values in \link{slcmControl}.}
}
\description{
Estimation for Parameters of \code{slcm} Object
}
