% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\alias{fit.slcm}
\title{Estimation for Parameters of \code{slcm} Object}
\usage{
fit(object, ...)

\method{fit}{slcm}(object, data,
    method = c("em", "hybrid", "nlm"),
    restriction = NULL,
    control = slcm.control(), ...)
}
\arguments{
\item{object}{a \code{slcm} object which defines latent structure to be estimated.}

\item{data}{a \code{data.frame} object which contains observed categorical variables beloning to the latent structure.}

\item{formula}{a \code{formula} object introducing exogenous covariates.}

\item{nrep}{number of trial}

\item{method}{estimating method for slcm parameters.}

\item{restriction}{a \code{list} of parameters to be restricted to zero.}

\item{control}{slcm control.}
}
\description{
Estimation for Parameters of \code{slcm} Object
}
