% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbering.R
\name{parameter_numbers}
\alias{parameter_numbers}
\title{Indexing \code{slcm} Parameters}
\usage{
parameter_numbers(object)
}
\arguments{
\item{object}{a \code{slcm} object}
}
\description{
For restriction of parameters,
}
