% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcm-wrapper.R
\name{print.slcm}
\alias{print.slcm}
\title{Print the SLCM object}
\usage{
\method{print}{slcm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{the \code{slcm} object.}

\item{digits}{the number of significant digits}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Print details and estimates found within the fitted SLCM.
Return the model invisibly (via \code{\link[=invisible]{invisible()}})
}
\description{
Custom printing class to reveal features of the fitted SLCM.
}
