% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regress}
\alias{regress}
\title{Regress Covariates on Latent Variables}
\usage{
regress(object, ...)
}
\arguments{
\item{object}{a slcm object}

\item{formula}{formula}

\item{data}{data}

\item{imputation}{imputation}

\item{method}{method}
}
\value{
coefficients
}
\description{
Regress Covariates on Latent Variables
}
