% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.slcm.R
\name{simulate.slcm}
\alias{simulate.slcm}
\title{Simulate \code{slcm} Object
Simulate data from \code{slcm} model.}
\usage{
\method{simulate}{slcm}(object, nsim = 500, nlevel = 2, random.param = TRUE, seed = NULL, ...)
}
\arguments{
\item{object}{a \code{slcm} object}

\item{nsim}{number of response data to simulate. Defaults to 100.}

\item{nlevel}{number of categories for each manifest item.}

\item{seed}{random seed}

\item{what}{objects to be printed among response, class, probs (params)}

\item{params}{parameters to be designated}
}
\description{
Simulate \code{slcm} Object
Simulate data from \code{slcm} model.
}
