% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcm-package.R
\docType{package}
\name{slcm-package}
\alias{slcm-package}
\title{slcm: Sparse Latent Class Model for Cognitive Diagnosis}
\description{
Perform a Bayesian estimation of the exploratory Sparse Latent Class Model for Binary Data described by Chen, Y., Culpepper, S. A., and Liang, F. (2020) \doi{10.1007/s11336-019-09693-2}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tmsalab/slcm}
  \item \url{https://tmsalab.github.io/slcm/}
  \item Report bugs at \url{https://github.com/tmsalab/slcm/issues}
}

}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{balamut2@illinois.edu} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Steven Andrew Culpepper \email{sculpepp@illinois.edu} (\href{https://orcid.org/0000-0003-4226-6176}{ORCID}) [copyright holder]
}

}
\keyword{internal}
