% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcm.R
\name{slcm}
\alias{slcm}
\title{Construct structural latent class model}
\usage{
slcm(x = NULL, ..., constraints = NULL)
}
\arguments{
\item{...}{additional formulae of the model}

\item{constraints}{a list containing constraints of the measurement invariance.}

\item{formula}{a formula for specifying latent structure. The details of model specification are below.}
}
\description{
Function for constructing multivariate latent class model structure
}
\details{
formula
}
