% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcmControl.R
\name{slcmControl}
\alias{slcmControl}
\title{Control Parameters for slcm Estimation}
\usage{
slcmControl(
  em.iterlim = 1000,
  em.tol = 1e-06,
  nlm.iterlim = 500,
  nlm.tol = 1e-08,
  init.param = NULL,
  nrep = 1,
  verbose = TRUE
)
}
\arguments{
\item{em.iterlim}{maximum number of iterations for EM algorithm. Default is 1000.}

\item{em.tol}{tolerance for the convergence of EM algorithm. Default is 1e-6.}

\item{nlm.iterlim}{maximum number of iterations for estimation with \code{nlm} function. Default is 500.}

\item{nlm.tol}{tolerance for the convergence of \code{nlm} function. Default is 1e-8.}

\item{init.param}{initial parameter}

\item{nrep}{number of trials}

\item{verbose}{a logical value whether to print estimation procedure.}
}
\value{
a \code{list} with control parameters for slcm estimation.
}
\description{
Control Parameters for slcm Estimation
}
\seealso{
\link{slcm}
}
