% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{compare}
\alias{compare}
\title{Comparing Two Estimated \code{slcm} Models}
\usage{
compare(
  model1,
  model2,
  test = c("chisq", "boot"),
  nboot = 100,
  maxiter = 500,
  tol = 1e-05,
  verbose = FALSE
)
}
\description{
Comparing Two Estimated \code{slcm} Models
}
