% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{confint.slcm}
\alias{confint.slcm}
\title{Confidence Intervals for Model Parameters}
\usage{
\method{confint}{slcm}(object, level = 0.95, type = c("param", "logit"), ...)
}
\arguments{
\item{object}{an object of class \code{slcm} and \code{estimated}.}

\item{level}{level a numeric value representing the desired confidence level for the intervals. Default is 0.95}

\item{type}{a character string specifying the format in which the results should be returned. Options are \code{"probs"} for probability format and \code{"logit"} for log-odds (logit) format. The default is \code{"probs"}.}
}
\description{
Computes confidence intervals for one or more parameters of fitted model. Package \pkg{slcm} adds methods for \code{slcm} fits.
}
