% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regress}
\alias{regress}
\title{Regress Exogenous Variables on Latent Variables}
\usage{
regress(object, ...)
}
\arguments{
\item{object}{an object of class \code{slcm} and \code{estimated}}

\item{formula}{a formula defining the regression model. It should include both latent class variables from the estimated model and any exogenous (external) variables}

\item{data}{an optional data frame containing the exogenous variables in interest.}

\item{imputation}{imputation method for latent class variables.}

\item{method}{bias-adjusting method for three-step approach}
}
\description{
This function allows you to perform regression analysis to understand the impact of exogenous (external) variables on the latent class variables in the estimated slcm model. Logistic regression and three-step approach is used for this purpose.
}
