% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.slcm.R
\name{simulate.slcm}
\alias{simulate.slcm}
\title{Simulate \code{slcm} Object
Simulate data from \code{slcm} model.}
\usage{
\method{simulate}{slcm}(object, nsim = 500, params = NULL, ncat = 2, seed, ...)
}
\arguments{
\item{object}{a \code{slcm} object}

\item{nsim}{number of response data to simulate. Defaults to 100.}

\item{params}{parameters to be designated}

\item{seed}{random seed}
}
\description{
Simulate \code{slcm} Object
Simulate data from \code{slcm} model.
}
