% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.slcm.R
\name{simulate.slcm}
\alias{simulate.slcm}
\title{Simulate \code{slcm} Object
Simulate data from \code{slcm} model.}
\usage{
\method{simulate}{slcm}(
  object,
  nobs = 500,
  nlevel = 2,
  what = c("response", "class", "probs"),
  params = NULL,
  seed,
  ...
)
}
\arguments{
\item{object}{a \code{slcm} object}

\item{params}{parameters to be designated}

\item{seed}{random seed}

\item{nsim}{number of response data to simulate. Defaults to 100.}
}
\description{
Simulate \code{slcm} Object
Simulate data from \code{slcm} model.
}
