% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slcmControl.R
\name{slcmControl}
\alias{slcmControl}
\title{Control Parameters for slcm Estimation}
\usage{
slcmControl(
  em.iterlim = 1000,
  em.tol = 1e-06,
  nlm.iterlim = 500,
  nlm.tol = 1e-08,
  init.param = NULL,
  nrep = 1,
  na.rm = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{em.iterlim}{an integer specifying maximum number of iterations allowed for EM algorithm. Default is 1000.}

\item{em.tol}{a numeric value setting tolerance for the convergence of EM algorithm. Default is 1e-6.}

\item{nlm.iterlim}{an integer specifying maximum number of iterations allowed for estimation with \code{nlm} function. Default is 500.}

\item{nlm.tol}{a numeric value setting tolerance for the convergence of \code{nlm} function. Default is 1e-8.}

\item{init.param}{initial parameters}

\item{nrep}{number of trials}

\item{na.rm}{a logical value indicating whether to remove missing values (NA)}

\item{verbose}{a logical value indicating whether to display progress updates during the estimation process}
}
\value{
a \code{list} with control parameters for slcm estimation.
}
\description{
Control Parameters for slcm Estimation
}
\seealso{
\link{slcm}
}
